Split up `GradleModuleMetadataWriter` into multiple files
Under the new `o.g.a.publish.internal.metadata` package.
